backend/requirements.txt
alembic==1.14.1
attrs==25.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.4
packaging==25.0
progressbar==2.5
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
